# Slightly less slow remote copyin