@geohot 2025-05-18
Nice! We still have the bounty for a tinygrad compatible parallel hash function, upped it to $300. "High performance crypto hash function in tinygrad making remote fast (max O(log(n)) kernels for len(n) message)"